feat: change some settings
