+++ .eslintrc.cjs
@@ -2,7 +2,7 @@ module.exports = {
     parser: '@typescript-eslint/parser',
     parserOptions: {
         project: 'tsconfig.json',
-        tsconfigRootDir : __dirname,
+        tsconfigRootDir: __dirname,
         sourceType: 'module',
     },
     plugins: ['@typescript-eslint/eslint-plugin'],
@@ -15,7 +15,7 @@ module.exports = {
         node: true,
         jest: true,
     },
-    ignorePatterns: ['.eslintrc.js', 'webpack.config.js', 'jest.config.js'],
+    ignorePatterns: ['.eslintrc.cjs', 'webpack.config.js', 'jest.config.js', 'front/main.js'],
     rules: {
         '@typescript-eslint/interface-name-prefix': 'off',
         '@typescript-eslint/explicit-function-return-type': 'off',
