+++ package.json
@@ -2,14 +2,15 @@
     "name": "ws_task",
     "version": "1.0.0",
     "description": "Template for the WebSocket remote control task",
-    "main": "index.js",
-    "type": "module",
+    "main": "index.ts",
     "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js",
-        "test": "echo \"Error: no test specified\" && exit 1"
+        "start": "node index.ts",
+        "start:dev": "nodemon --exec ts-node-dev --respawn --transpile-only index.ts",
+        "test": "jest",
+        "lint": "eslint src/http_server/server.ts",
+        "lint:fix": "eslint ./src --fix"
     },
-    "author": "Andrei Auchynnikau",
+    "author": "Mikalai Rusakovich",
     "keywords": [],
     "license": "ISC",
     "devDependencies": {
@@ -17,6 +18,7 @@
         "@types/node": "^20.11.19",
         "@types/supertest": "^6.0.2",
         "@types/uuid": "^9.0.8",
+        "@types/ws": "^8.5.10",
         "@typescript-eslint/eslint-plugin": "^7.0.1",
         "@typescript-eslint/parser": "^7.0.1",
         "cross-env": "^7.0.3",
